fix : 타입 선언 리팩토링
